Scrap old channel interfaces and attempt to unify them
Offer a method to write bytes, which is what we get when dealing with
encryption and base64 encodings. For some reason, a byte array is
cumbersome to convert to a char array in Java.

We always need to read String lines; everything we receive should
contain only valid ASCII characters since we will base64 encode all of
it.