Repair client into a compilable state
This is our new client structure, which will hopefully make things
much simpler for us. We have a NetworkListenerThread, which is only
responsible for reading lines from the socket; an input loop,
responsible for reading lines input by the user; and a ProcessorThread,
which handles all the actual work, fueled by a BlockingQueue.