Preparations for disconnection handling
Unfortunately, this has turned into a rather confused and unwieldy
commit. Several things are going on here; Parcels have been split into
Parcels (just a type), StringParcels (a message), and SocketParcels(a
socket).

The NetworkListenerThread now tries to establish a connection
periodically until it is interrupted. When a connection is established,
it sends the new socket to the processor thread. When it is lost, it
also sends a parcel.

The current state is still missing some vital functionality, like
setting the in the processor thread.